fix(logger-avr): fix printed ID compiled with avr-gcc
This issue happened on the Arduino Uno. The ID stopped after
the first bracket. This is because the variable width precision
in snprintf (asteris) is not supported on avr-gcc. This fix separates the
generation of the format string. Idea: Perform this on construction,
as the format does not depend on dynamic
variables.